fix(profile): avoid build-time prerender errors by wrapping client page in server wrapper with dynamic rendering
