+++ app/profile/ProfileClient.tsx
@@ -1,0 +1,105 @@
+'use client'
+
+import { Button } from '@/components/ui/button'
+import { Input } from '@/components/ui/input'
+import { Label } from '@/components/ui/label'
+import { Separator } from '@/components/ui/separator'
+import { createClient } from '@/lib/supabase/client'
+import { useEffect, useState } from 'react'
+import { User } from '@supabase/supabase-js'
+import { LogoutButton } from '@/components/auth/logout-button'
+
+export default function ProfileClient() {
+  const supabase = createClient()
+  const [loading, setLoading] = useState(true)
+  const [user, setUser] = useState<User | null>(null)
+  const [name, setName] = useState('')
+  const [email, setEmail] = useState('')
+
+  useEffect(() => {
+    const fetchProfile = async () => {
+      setLoading(true)
+      const {
+        data: { user },
+      } = await supabase.auth.getUser()
+
+      if (user) {
+        setUser(user)
+        setEmail(user.email || '')
+        const { data: userData, error } = await supabase
+          .from('users')
+          .select('name')
+          .eq('id', user.id)
+          .single()
+
+        if (error) {
+          console.error('Error fetching user name', error)
+        } else if (userData) {
+          setName(userData.name || '')
+        }
+      }
+      setLoading(false)
+    }
+
+    fetchProfile()
+  }, [supabase])
+
+  const [saving, setSaving] = useState(false)
+
+  const handleSave = async () => {
+    if (!user) return
+    setSaving(true)
+    const { error } = await supabase
+      .from('users')
+      .update({ name: name })
+      .eq('id', user.id)
+
+    if (error) {
+      alert('Error updating profile: ' + error.message)
+    } else {
+      alert('Profile saved successfully!')
+    }
+    setSaving(false)
+  }
+
+  if (loading) {
+    return (
+      <div className="flex-1 space-y-4 p-4 md:p-8 pt-6">
+        <p>Loading...</p>
+      </div>
+    )
+  }
+
+  return (
+    <div className="flex-1 space-y-4 p-4 md:p-8 pt-6">
+      <div className="flex items-center justify-between space-y-2">
+        <h2 className="text-3xl font-bold tracking-tight">Profile</h2>
+      </div>
+      <Separator />
+      <div className="space-y-8 max-w-2xl mx-auto py-8">
+        <div className="grid gap-2">
+          <Label htmlFor="email">Email</Label>
+          <Input id="email" type="email" value={email} disabled />
+        </div>
+
+        <div className="grid gap-2">
+          <Label htmlFor="name">Name</Label>
+          <Input
+            id="name"
+            type="text"
+            value={name}
+            onChange={(e) => setName(e.target.value)}
+          />
+        </div>
+
+        <div className="flex space-x-2">
+          <Button onClick={handleSave} disabled={saving}>
+            {saving ? 'Saving...' : 'Save Changes'}
+          </Button>
+          <LogoutButton />
+        </div>
+      </div>
+    </div>
+  )
+}
+
